requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
coverage==7.0.5
Django==4.1.7
django-crispy-forms==1.12.0
django-filter==22.1
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
